Parenthesis fixing
Fixing bugs where the parenthesis where not being resolved correctly.  This is the end of the work that I will be putting in on this code for the near future.  I may over break make this work for double's for fun.  But the app as it is is sufficient  for this classes purposes.